pyproject.toml
[project]
name = "Dev10x"
version = "0.55.0"
description = "Claude Code plugin providing reusable skills, hooks, and commands"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "msgpack>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
dev10x = "dev10x.cli:cli"

[project.optional-dependencies]
dev = [
    "factory-boy>=3.3",
    "faker>=33.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-benchmark>=5.0",
    "pytest-cov>=6.0",
    "mcp>=1.0",
    "ruff>=0.9",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
norecursedirs = ["codex-skills"]
testpaths = ["tests"]
addopts = "--cov=src/dev10x --cov-report=term-missing"

[tool.coverage.run]
source = ["src/dev10x"]

[tool.coverage.report]
fail_under = 38
show_missing = true
exclude_lines = ["if __name__", "pragma: no cover"]

[tool.ruff]
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
mypy_path = "src"
